+++ crates/libafl_structured_mutators/src/mutators/integers.rs
@@ -305,8 +305,8 @@ macro_rules! impl_int_default_mutator {
         where
             S: libafl::state::HasRand + core::fmt::Debug,
         {
-            fn weight(&self, _data: & $name) -> u64 {
-                self.mutations.len() as u64
+            fn weight(&self, _data: & $name) -> usize {
+                self.mutations.len()
             }
 
             fn mutate(&mut self, data: &mut $name, state: &mut S) -> bool {
